Refactor WordList methods
