+++ OfficeIMO.Word/WordList.cs
@@ -34,44 +34,6 @@ public partial class WordList : WordElement {
     /// <summary>
     /// Gets all the list items associated with this WordList.
     /// </summary>
-    //private string NsidId {
-    //    get {
-    //        if (AbstractNum == null) {
-    //            return null;
-    //        }
-
-    //        return AbstractNum.Nsid.Val;
-
-    //    }
-    //    set {
-    //        if (AbstractNum != null) {
-    //            AbstractNum.Nsid.Val = value;
-    //        }
-    //    }
-    //}
-
-    //private string GenerateNsidId() {
-    //    // https://learn.microsoft.com/en-us/dotnet/api/documentformat.openxml.wordprocessing.nsid?view=openxml-2.8.1
-    //    // Specifies a number value specified as a four digit hexadecimal number),
-    //    // whose contents of this decimal number are interpreted based on the context of the parent XML element.
-    //    // for example FFFFFF89 or D9842532
-    //    return Guid.NewGuid().ToString().ToUpper().Substring(0, 8);
-
-    //}
-
-    //private AbstractNum AbstractNum {
-    //    get {
-    //        var numbering = _document._wordprocessingDocument.MainDocumentPart!.NumberingDefinitionsPart!.Numbering;
-    //        var abstractNumList = numbering.ChildElements.OfType<AbstractNum>();
-    //        foreach (var abstractNum in abstractNumList) {
-    //            if (abstractNum.AbstractNumberId == _abstractId) {
-    //                return abstractNum;
-    //            }
-    //        }
-
-    //        return null;
-    //    }
-    //}
     public List<WordParagraph> ListItems {
         get {
             List<WordParagraph> list = new List<WordParagraph>();
@@ -453,375 +415,4 @@ public partial class WordList : WordElement {
         _headerFooter = headerFooter;
     }
 
-    /// <summary>
-    /// Adds an item to the list using an existing paragraph.
-    /// </summary>
-    /// <param name="wordParagraph">The paragraph to add.</param>
-    /// <param name="level">The list level.</param>
-    /// <returns>The added <see cref="WordParagraph"/>.</returns>
-    public WordParagraph AddItem(WordParagraph wordParagraph, int level = 0) {
-        return AddItem(null, level, wordParagraph);
-    }
-
-    /// <summary>
-    /// Adds an item to the list with specified text.
-    /// </summary>
-    /// <param name="text">The text of the list item.</param>
-    /// <param name="level">The list level.</param>
-    /// <param name="wordParagraph">An optional existing paragraph.</param>
-    /// <returns>The added <see cref="WordParagraph"/>.</returns>
-    public WordParagraph AddItem(string text, int level = 0, WordParagraph wordParagraph = null) {
-        var paragraph = new Paragraph();
-        var run = new Run();
-        run.Append(new RunProperties());
-        run.Append(new Text { Space = SpaceProcessingModeValues.Preserve });
-
-        var paragraphProperties = new ParagraphProperties();
-        paragraphProperties.Append(new ParagraphStyleId { Val = "ListParagraph" });
-        paragraphProperties.Append(
-            new NumberingProperties(
-                new NumberingLevelReference { Val = level },
-                new NumberingId { Val = _numberId }
-            ));
-        paragraph.Append(paragraphProperties);
-        paragraph.Append(run);
-
-        // Determine proper placement for the paragraph
-        if (wordParagraph != null) {
-            // If a specific paragraph reference is provided, insert after it
-            wordParagraph._paragraph.InsertAfterSelf(paragraph);
-        } else if (this.ListItems.Count == 0 && _wordParagraph != null) {
-            // First item in a paragraph-referenced list - insert after reference paragraph
-            _wordParagraph._paragraph.InsertAfterSelf(paragraph);
-        } else if (_isToc || IsToc) {
-            // TOC list items should be placed at the end of the document
-            _wordprocessingDocument.MainDocumentPart!.Document.Body!.AppendChild(paragraph);
-        } else if (_headerFooter != null) {
-            // Header/footer list items
-            if (_headerFooter._header != null) {
-                _headerFooter._header.Append(paragraph);
-            } else if (_headerFooter._footer != null) {
-                _headerFooter._footer.Append(paragraph);
-            }
-        } else if (_wordParagraph != null && _wordParagraph._paragraph.Parent is TableCell) {
-            // Handle table cell lists
-            var parent = _wordParagraph._paragraph.Parent;
-            if (this.ListItems.Count > 0) {
-                var lastItem = this.ListItems.Last();
-                lastItem._paragraph.InsertAfterSelf(paragraph);
-            } else {
-                parent.Append(paragraph);
-            }
-        } else {
-            // For standard lists without specific placement, add at the end
-            _wordprocessingDocument.MainDocumentPart!.Document.Body!.AppendChild(paragraph);
-        }
-
-        var newParagraph = new WordParagraph(_document, paragraph, run);
-        if (text != null) {
-            newParagraph.Text = text;
-        }
-
-        // Handle TOC styling
-        if (_isToc || IsToc) {
-            newParagraph.Style = WordParagraphStyle.GetStyle(level);
-        }
-
-        return newParagraph;
-    }
-
-    /// <summary>
-    /// Removes the list and its items from the document.
-    /// </summary>
-    public void Remove() {
-        var numberingPart = _document._wordprocessingDocument.MainDocumentPart.NumberingDefinitionsPart;
-        var numbering = numberingPart?.Numbering;
-        if (numbering != null) {
-            var abstractNum = numbering.Elements<AbstractNum>().FirstOrDefault(a => a.AbstractNumberId.Value == _abstractId);
-            abstractNum?.Remove();
-
-            var numberingInstance = numbering.Elements<NumberingInstance>().FirstOrDefault(n => n.NumberID.Value == _numberId);
-            numberingInstance?.Remove();
-
-            if (!numbering.ChildElements.OfType<AbstractNum>().Any() &&
-                !numbering.ChildElements.OfType<NumberingInstance>().Any() &&
-                !numbering.ChildElements.OfType<NumberingPictureBullet>().Any()) {
-                _document._wordprocessingDocument.MainDocumentPart.DeletePart(numberingPart);
-            }
-        }
-
-        foreach (var listItem in ListItems.ToList()) {
-            listItem.Remove();
-        }
-    }
-
-    /// <summary>
-    /// Creates a deep copy of this list and inserts it after the last item of the list.
-    /// </summary>
-    /// <returns>The cloned <see cref="WordList"/>.</returns>
-    /// <example>
-    /// <code>
-    /// var list = document.AddList(WordListStyle.Bulleted);
-    /// list.AddItem("First");
-    /// var duplicate = list.Clone();
-    /// </code>
-    /// </example>
-    public WordList Clone() {
-        var reference = ListItems.LastOrDefault()?._paragraph;
-        if (reference == null) {
-            throw new InvalidOperationException("Cannot clone an empty list.");
-        }
-        return Clone(reference, true);
-    }
-
-    /// <summary>
-    /// Clones the list and inserts it relative to the specified paragraph.
-    /// </summary>
-    /// <param name="paragraph">Reference paragraph for insertion.</param>
-    /// <param name="after">If true inserts after the paragraph, otherwise before.</param>
-    /// <returns>The cloned <see cref="WordList"/>.</returns>
-    /// <example>
-    /// <code>
-    /// var duplicate = list.Clone(paragraph, after: false);
-    /// </code>
-    /// </example>
-    public WordList Clone(WordParagraph paragraph, bool after = true) {
-        return Clone(paragraph._paragraph, after);
-    }
-
-    private WordList Clone(OpenXmlElement referenceParagraph, bool after) {
-        var numberingPart = _document._wordprocessingDocument.MainDocumentPart.NumberingDefinitionsPart;
-        if (numberingPart == null) {
-            numberingPart = _document._wordprocessingDocument.MainDocumentPart.AddNewPart<NumberingDefinitionsPart>();
-            numberingPart.Numbering = new Numbering();
-        }
-        var numbering = numberingPart.Numbering;
-
-        var originalAbstract = numbering.Elements<AbstractNum>().First(a => a.AbstractNumberId.Value == _abstractId);
-        var originalInstance = numbering.Elements<NumberingInstance>().First(n => n.NumberID.Value == _numberId);
-
-        int newAbstractId = GetNextAbstractNum(numbering);
-        int newNumberId = GetNextNumberingInstance(numbering);
-
-        var newAbstract = (AbstractNum)originalAbstract.CloneNode(true);
-        newAbstract.AbstractNumberId = newAbstractId;
-        numbering.Append(newAbstract);
-
-        var newInstance = (NumberingInstance)originalInstance.CloneNode(true);
-        newInstance.NumberID = newNumberId;
-        var absId = newInstance.GetFirstChild<AbstractNumId>();
-        if (absId != null) absId.Val = newAbstractId;
-        numbering.Append(newInstance);
-
-        WordList clonedList;
-        if (_headerFooter != null) {
-            clonedList = new WordList(_document, _headerFooter);
-        } else if (_wordParagraph != null) {
-            clonedList = new WordList(_document, _wordParagraph, _isToc);
-        } else {
-            clonedList = new WordList(_document, _isToc);
-        }
-
-        clonedList._abstractId = newAbstractId;
-        clonedList._numberId = newNumberId;
-
-        OpenXmlElement currentRef = referenceParagraph;
-        WordParagraph firstInserted = null;
-        foreach (var item in ListItems) {
-            var clonedParagraph = (Paragraph)item._paragraph.CloneNode(true);
-            var numberingProps = clonedParagraph.GetFirstChild<ParagraphProperties>()?.NumberingProperties;
-            if (numberingProps != null) {
-                numberingProps.NumberingId.Val = newNumberId;
-            }
-            currentRef = after ? currentRef.InsertAfterSelf(clonedParagraph) : currentRef.InsertBeforeSelf(clonedParagraph);
-            if (firstInserted == null) {
-                firstInserted = new WordParagraph(_document, (Paragraph)currentRef);
-            }
-        }
-
-        if (firstInserted != null) {
-            clonedList._wordParagraph = firstInserted;
-        }
-
-        return clonedList;
-    }
-
-    /// <summary>
-    /// Merges another list into this list.
-    /// </summary>
-    /// <param name="documentList">The list to merge.</param>
-    public void Merge(WordList documentList) {
-        // Reattach all items from the other list to this list
-        foreach (var item in documentList.ListItems) {
-            var numberingProperties = item._paragraphProperties.NumberingProperties;
-            // Change the NumId to the NumId of this list
-            if (numberingProperties != null && numberingProperties.NumberingId != null) {
-                numberingProperties.NumberingId.Val = this._numberId;
-            }
-        }
-        // Remove the other list
-        documentList.Remove();
-    }
-
-    /// <summary>
-    /// Adds a custom bullet list with a single level using provided formatting.
-    /// </summary>
-    /// <param name="document">Target document.</param>
-    /// <param name="symbol">Bullet symbol.</param>
-    /// <param name="fontName">Font name for the symbol.</param>
-    /// <param name="colorHex">Hex color of the symbol.</param>
-    /// <param name="fontSize">Font size in points.</param>
-    /// <returns>The created <see cref="WordList"/>.</returns>
-    public static WordList AddCustomBulletList(WordDocument document, char symbol, string fontName, string colorHex, int? fontSize = null) {
-        if (document == null) throw new ArgumentNullException(nameof(document));
-
-        var list = document.AddList(WordListStyle.Custom);
-
-        var level = new Level();
-        level.Append(new StartNumberingValue() { Val = 1 });
-        level.Append(new NumberingFormat() { Val = NumberFormatValues.Bullet });
-        level.Append(new LevelText() { Val = symbol.ToString() });
-        level.Append(new LevelJustification() { Val = LevelJustificationValues.Left });
-
-        var prevProps = new PreviousParagraphProperties();
-        prevProps.Append(new Indentation() { Left = "720", Hanging = "360" });
-        level.Append(prevProps);
-
-        var symbolProps = new NumberingSymbolRunProperties();
-        if (!string.IsNullOrEmpty(fontName)) {
-            symbolProps.Append(new RunFonts { Ascii = fontName, HighAnsi = fontName });
-        }
-        if (!string.IsNullOrEmpty(colorHex)) {
-            symbolProps.Append(new DocumentFormat.OpenXml.Wordprocessing.Color { Val = colorHex.Replace("#", "").ToLowerInvariant() });
-        }
-        if (fontSize.HasValue) {
-            var size = (fontSize.Value * 2).ToString();
-            symbolProps.Append(new FontSize { Val = size });
-            symbolProps.Append(new FontSizeComplexScript { Val = size });
-        }
-        level.Append(symbolProps);
-
-        list.Numbering.AddLevel(level);
-
-        return list;
-    }
-
-    /// <summary>
-    /// Adds a custom bullet list with a single level using provided formatting.
-    /// </summary>
-    /// <param name="document">Target document.</param>
-    /// <param name="symbol">Bullet symbol.</param>
-    /// <param name="fontName">Font name for the symbol.</param>
-    /// <param name="color">Color of the symbol.</param>
-    /// <param name="colorHex">Hex color fallback.</param>
-    /// <param name="fontSize">Font size in points.</param>
-    /// <returns>The created <see cref="WordList"/>.</returns>
-    public static WordList AddCustomBulletList(WordDocument document, WordBulletSymbol symbol, string fontName, SixLabors.ImageSharp.Color? color = null, string colorHex = null, int? fontSize = null) {
-        string finalColor = color?.ToHexColor() ?? colorHex;
-        return AddCustomBulletList(document, (char)symbol, fontName, finalColor, fontSize);
-    }
-
-    /// <summary>
-    /// Adds a custom list with no predefined levels so that levels can be configured manually.
-    /// </summary>
-    /// <param name="document">Target document.</param>
-    /// <returns>The created <see cref="WordList"/>.</returns>
-    public static WordList AddCustomList(WordDocument document) {
-        if (document == null) throw new ArgumentNullException(nameof(document));
-
-        var list = document.AddList(WordListStyle.Custom);
-        list.Numbering.RemoveAllLevels();
-        return list;
-    }
-
-    /// <summary>
-    /// Adds a list level using the provided formatting. If there are gaps in the level index,
-    /// missing levels are created using the formatting of the previously added level.
-    /// </summary>
-    /// <param name="levelIndex">1-based level index.</param>
-    /// <param name="symbol">Bullet symbol.</param>
-    /// <param name="fontName">Font name.</param>
-    /// <param name="colorHex">Hex color.</param>
-    /// <param name="fontSize">Font size in points.</param>
-    /// <returns>The <see cref="WordList"/> for chaining.</returns>
-    public WordList AddListLevel(int levelIndex, char symbol, string fontName, string colorHex, int? fontSize = null) {
-        if (levelIndex < 1) throw new ArgumentOutOfRangeException(nameof(levelIndex));
-
-        var currentCount = this.Numbering.Levels.Count;
-        if (currentCount > 0 && levelIndex > currentCount + 1) {
-            var last = this.Numbering.Levels.Last()._level;
-            while (this.Numbering.Levels.Count < levelIndex - 1) {
-                var clone = (Level)last.CloneNode(true);
-                this.Numbering.AddLevel(clone);
-            }
-        }
-
-        if (levelIndex > 1 && this.Numbering.Levels.Count == 0) {
-            levelIndex = 1;
-        }
-
-        var newLevel = CreateBulletLevel(symbol, fontName, colorHex, fontSize);
-        this.Numbering.AddLevel(newLevel);
-        return this;
-    }
-
-    /// <summary>
-    /// Adds a list level using the provided formatting.
-    /// </summary>
-    /// <param name="levelIndex">1-based level index.</param>
-    /// <param name="symbol">Bullet symbol.</param>
-    /// <param name="fontName">Font name.</param>
-    /// <param name="color">Color for bullet.</param>
-    /// <param name="colorHex">Hex color fallback.</param>
-    /// <param name="fontSize">Font size in points.</param>
-    /// <returns>The <see cref="WordList"/> for chaining.</returns>
-    public WordList AddListLevel(int levelIndex, WordBulletSymbol symbol, string fontName, SixLabors.ImageSharp.Color? color = null, string colorHex = null, int? fontSize = null) {
-        string finalColor = color?.ToHexColor() ?? colorHex;
-        return AddListLevel(levelIndex, (char)symbol, fontName, finalColor, fontSize);
-    }
-
-    private static Level CreateBulletLevel(char symbol, string fontName, string colorHex, int? fontSize) {
-        var level = new Level();
-        level.Append(new StartNumberingValue() { Val = 1 });
-        level.Append(new NumberingFormat() { Val = NumberFormatValues.Bullet });
-        level.Append(new LevelText() { Val = symbol.ToString() });
-        level.Append(new LevelJustification() { Val = LevelJustificationValues.Left });
-
-        var prevProps = new PreviousParagraphProperties();
-        prevProps.Append(new Indentation() { Left = "720", Hanging = "360" });
-        level.Append(prevProps);
-
-        var symbolProps = new NumberingSymbolRunProperties();
-        if (!string.IsNullOrEmpty(fontName)) {
-            symbolProps.Append(new RunFonts { Ascii = fontName, HighAnsi = fontName });
-        }
-        if (!string.IsNullOrEmpty(colorHex)) {
-            symbolProps.Append(new DocumentFormat.OpenXml.Wordprocessing.Color { Val = colorHex.Replace("#", "").ToLowerInvariant() });
-        }
-        if (fontSize.HasValue) {
-            var size = (fontSize.Value * 2).ToString();
-            symbolProps.Append(new FontSize { Val = size });
-            symbolProps.Append(new FontSizeComplexScript { Val = size });
-        }
-        level.Append(symbolProps);
-        return level;
-    }
-
-    private static void EnsureW15Namespace(Numbering numbering) {
-        const string prefix = "w15";
-        const string ns = "http://schemas.microsoft.com/office/word/2012/wordml";
-        if (numbering.LookupNamespace(prefix) == null) {
-            numbering.AddNamespaceDeclaration(prefix, ns);
-        }
-        if (numbering.MCAttributes == null) {
-                numbering.MCAttributes = new MarkupCompatibilityAttributes { Ignorable = prefix };
-        } else {
-            var ignorable = numbering.MCAttributes.Ignorable?.Value;
-            if (string.IsNullOrEmpty(ignorable)) {
-                numbering.MCAttributes.Ignorable = prefix;
-            } else if (!ignorable.Split(new[] { ' ' }, StringSplitOptions.RemoveEmptyEntries).Contains(prefix)) {
-                numbering.MCAttributes.Ignorable = ignorable + " " + prefix;
-            }
-        }
-    }
 }
